% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_home_lookup_function.r
\name{build.home.lookup}
\alias{build.home.lookup}
\title{Function to build the home growth lookup table
called by ZPDGrowthTrajectories() function}
\usage{
build.home.lookup(
  integration.points,
  ZPD.width,
  ZPD.offset,
  rate,
  maxachievement
)
}
\arguments{
\item{integration.points}{integration points for numerical integration}

\item{ZPD.width}{the radius of the student's ZPD}

\item{ZPD.offset}{offset of the student's ZPD from the current level of achievement}

\item{rate}{exponential decay rate for the home curriculum function}

\item{maxachievement}{the maximum achieveable level of achievement}
}
\description{
Function to build the home growth lookup table
called by ZPDGrowthTrajectories() function
}
\examples{
build.home.lookup(integration.points=20, ZPD.width=.05, ZPD.offset=.02, rate=4, maxachievement=1.5)
}
