% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_school_lookup_function.r
\name{build.school.lookup}
\alias{build.school.lookup}
\title{Function to build the school growth lookup table
called by ZPDGrowthTrajectories() function}
\usage{
build.school.lookup(
  integration.points,
  ZPD.width,
  ZPD.offset,
  curriculum.start.points,
  curriculum.widths,
  slope1,
  slope2,
  maxachievement
)
}
\arguments{
\item{integration.points}{integration points for numerical integration}

\item{ZPD.width}{the radius of the student's ZPD}

\item{ZPD.offset}{offset of the student's ZPD from the current level of achievement}

\item{curriculum.start.points}{matrix or list of matrices of curriculum starting locations}

\item{curriculum.widths}{matrix or list of matrices of curriculum widths}

\item{slope1}{the slope of the lower leg of the school curriculum}

\item{slope2}{the slope of the upper leg of the school curriculum}

\item{maxachievement}{the maximum achieveable level of achievement}
}
\description{
Function to build the school growth lookup table
called by ZPDGrowthTrajectories() function
}
\examples{
\dontrun{
curriculum.start.points <- matrix(c(.1, .2, .3), ncol=1)
curriculum.widths <- matrx(rep(.11, 3), ncol=1)
build.school.lookup(integration.points=20, ZPD.width=.05, ZPD.offset=.02,
                      curriculum.start.points=curriculum.start.points,
                      curriculum.widths=curriculum.widths,
                      slope1=10, slope2=30, maxachievement=1.5)
 }
}
