% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_curriculum_function.R
\name{buildHomeCurriculum}
\alias{buildHomeCurriculum}
\title{A helper function for describing the home curriculum.}
\usage{
buildHomeCurriculum(points, shape1, shape2)
}
\arguments{
\item{points}{an integer number of integration points. Higher means more accuracy but slower execution.}

\item{shape1}{shape1 parameter for \code{dbeta}.}

\item{shape2}{shape2 parameter for \code{dbeta}.}
}
\description{
\code{buildHomeCurriculum} creates the home curriculum object expected by \code{ZPDGrowthTrajectories}.
}
\details{
This function creates the home curriculum in the format expected by the
expected by the \code{ZPDGrowthTrajectories} function.
It is an internal function and not intended for stand-alone use.
The home curriculum is defined according to a beta probability
density function defined on [0,1].
}
