% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_curriculum_function.R
\name{buildSchoolCurriculum}
\alias{buildSchoolCurriculum}
\title{A helper function for describing the school curriculum.}
\usage{
buildSchoolCurriculum(start_point, width, assignment, points, n1, n3, alpha)
}
\arguments{
\item{start_point}{a matrix providing the start points of each grade level (rows) and version (columns) of the
school curriculum.}

\item{width}{a matrix providing the spans of each grade level (rows) and version (columns) of the
school curriculum.}

\item{assignment}{a vector. The length is the number of days to simulate. Each entry contains a number representing which
grade-level curriculum to present. Zero denotes summers. The numbers correspond to the row index of the
 \code{start_point} and \code{width} objects.}

\item{points}{number of integration points. Controls tradeoff between accuracy and execution speed.
Defaults to 200.}

\item{n1}{scalar argument to the \code{n1} argument of \code{trapedoid::dtrapezoid}; controls
steepness of the school curriculum cutoff at the lower range. Conceptually controls the amount of review content.}

\item{n3}{scalar argument to the \code{n3} argument of \code{trapedoid::dtrapezoid}; controls
steepness of the school curriculum cutoff at the upper range. Conceptually controls the amount of advanced content.}

\item{alpha}{scalar argument to the \code{alpha} argument of \code{trapedoid::dtrapezoid}. Defaults to 1.}
}
\description{
\code{buildSchoolCurriculum} creates the school curriculum object expected by \code{ZPDGrowthTrajectories}.
}
\details{
This function takes the matrix of start points, widths, and assignment,
 as well as the control parameters min, max, n1, n3, and alpha, and
 creates a list mapping the parameters of the trapezoidal probability density function
 to each grade level ("index"). Index=0 refers to periods of no school (e.g., summers).
 It is an internal function and not intended for stand-alone use.
}
