% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeTrajectories_function.r
\name{describeTrajectories}
\alias{describeTrajectories}
\title{Create summary statistics for ZPDGrowthTrajectories}
\usage{
describeTrajectories(trajectories, assignment = NULL, byCurriculum = TRUE)
}
\arguments{
\item{trajectories}{An object of class \code{ZPD} produced by the \code{ZPDGrowthTrajectories()}
function. If needed, this object be converted internally to "long" format.}

\item{assignment}{a vector indicating which school curriculum, if any, was provided during
each time interval. If provided, summary statistics are calculated at each transition. Defaults
to NULL.}

\item{byCurriculum}{Logical. Should descriptives be broken down by curriculum? Defaults
to TRUE.}
}
\description{
\code{describeTrajectories} calculates summary statistics by time interval and curriculum
for \code{ZPDGrowthTrajectories} output.
}
