% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_trajectories_function.R
\name{growTrajectories}
\alias{growTrajectories}
\title{Function for creating growth trajectories.}
\usage{
growTrajectories(days, points, learn.rate, decay.rate, initial.ach, home.env,
  dose, decay.weight, school.curr.fcn.values.by.day, home.curr.fcn.values,
  zpd.offset, zpd.sd, zpd.scale, useGPU = FALSE)
}
\arguments{
\item{days}{the number of days to simulate.}

\item{points}{the number of integration points.}

\item{learn.rate}{vector of learning rates, one per child.}

\item{decay.rate}{vector of decay rates, one per child. Controls forgetting.}

\item{initial.ach}{vector of initial achievements, one per child.}

\item{home.env}{vector of home environments, one per child.}

\item{dose}{scalar dose parameter, controls mixing of school curriculum and home curriculum during school years. Must be [0,1].}

\item{decay.weight}{scalar parameter, global control of decay rates.}

\item{school.curr.fcn.values.by.day}{object produced by the \code{buildSchoolCurriculum} function.}

\item{home.curr.fcn.values}{object produced by the \code{buildHomeCurriculum} function.}

\item{zpd.offset}{scalar value, measured on the same scale as achievement, describing where the ZPD peaks
relative to the current achievement.}

\item{zpd.sd}{the standard deviation of the ZPD function, controls its width.}

\item{zpd.scale}{scaling factor for the ZPD function, controls overall learning rate.}

\item{useGPU}{should the code be executed on the GPU? May offer performance boost on large simulations. The \code{gpuR}
package must be installed.}
}
\description{
\code{growTrajectories} calculates achievement at each day.
}
\details{
This function is called by \code{ZPDGrowthTrajectories} to iteratively calculate achievement for each day.
}
