% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_growth_rate_function.r
\name{home.growth.rate}
\alias{home.growth.rate}
\title{Function for calculating growth resulting from home experience at every possible achievement level,
 given a level of achievement, ZPD parameters, and home curriculum parameters.
 This is used to populate the lookup table for home growth.
Called by build.home.lookup()}
\usage{
home.growth.rate(achievement, ZPD.offset, ZPD.width, rate, integration.points)
}
\arguments{
\item{achievement}{a vector of achievement levels}

\item{ZPD.offset}{offset of the student's ZPD from the current level of achievement}

\item{ZPD.width}{the radius of the student's ZPD}

\item{rate}{exponential decay rate for the home curriculum function}

\item{integration.points}{integration points for numerical integration}
}
\description{
Function for calculating growth resulting from home experience at every possible achievement level,
 given a level of achievement, ZPD parameters, and home curriculum parameters.
 This is used to populate the lookup table for home growth.
Called by build.home.lookup()
}
\examples{
\dontrun{
home.growth.rate(achievement=seq(0,1, .1), integration.points=200, ZPD.offset=0,
                 ZPD.width=.04, rate=5)
}
}
