% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_growth_function.r
\name{home.growth}
\alias{home.growth}
\title{Function for calculating the growth during interval t from exposure to home
called by home.growth.rate() function}
\usage{
home.growth(x, achievement, ZPD.offset, ZPD.width, rate)
}
\arguments{
\item{x}{achievement value}

\item{achievement}{the student's current achievement}

\item{ZPD.offset}{offset of the student's ZPD from the current level of achievement}

\item{ZPD.width}{the radius of the student's ZPD}

\item{rate}{exponential decay rate for the home curriculum function}
}
\description{
Function for calculating the growth during interval t from exposure to home
called by home.growth.rate() function
}
\examples{
\dontrun{
home.growth(x=.2, achievement=.2, ZPD.offset=0, ZPD.width=.1, rate=5)
}
}
