% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_function.r
\name{home}
\alias{home}
\title{Function to define the home curriculum function
Using exponential decay
called by home.growth() function
Home curriculum intensity given by (-rate*x)}
\usage{
home(x, rate)
}
\arguments{
\item{x}{level of achievement}

\item{rate}{exponential decay rate}
}
\description{
Function to define the home curriculum function
Using exponential decay
called by home.growth() function
Home curriculum intensity given by (-rate*x)
}
\examples{
x <- seq(.0001, 1, .001)
plot(x, y=home(x, rate=5), "l")
}
