% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.melt_function.R
\name{matrix.melt}
\alias{matrix.melt}
\title{Function for restructuring matrices from wide to long.}
\usage{
matrix.melt(matrix)
}
\arguments{
\item{matrix}{Matrix to be transposed.}
}
\description{
\code{matrix.melt()} restructures matrices from wide to long.
}
\details{
This function restructures matrices. It works similarly to reshape2::melt but does not coerce
  matrices to data frames. Used interally by \code{ZPDGrowthTrajectories()} and not suited for general
  use. 
  
  This code assumes that the first column contains student IDs. These are dropped.
}
