% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_achievement_function.r
\name{max.achievement}
\alias{max.achievement}
\title{Function for calculating the max value that should be included in the lookup table
based on the max school curriculum and the point at which the home growth rate
falls below some negligible threshold
called by ZPDGrowthTrajectories() function}
\usage{
\method{max}{achievement}(
  ZPD.width,
  ZPD.offset,
  curriculum.start.points,
  curriculum.widths,
  slope2,
  rate,
  threshold,
  integration.points
)
}
\arguments{
\item{ZPD.width}{the radius of the student's ZPD}

\item{ZPD.offset}{offset of the student's ZPD from the current level of achievement}

\item{curriculum.start.points}{matrix of curriculum starting locations}

\item{curriculum.widths}{matrix of curriculum widths}

\item{slope2}{matrix of slopes of the upper leg of the school curriculum}

\item{rate}{exponential decay rate for the home curriculum function}

\item{threshold}{the threshold growth rate for finding the effective maximum possible achievement}

\item{integration.points}{integration points for numerical integration}
}
\description{
Function for calculating the max value that should be included in the lookup table
based on the max school curriculum and the point at which the home growth rate
falls below some negligible threshold
called by ZPDGrowthTrajectories() function
}
\examples{
\dontrun{
curriculum.start.points <- matrix(c(.1, .2, .3), ncol=1)
curriculum.widths <- matrix(rep(.11, 3), ncol=1)
slope2 <- matrix(rep(10, 3), ncol=1)

max.achievement(ZPD.width=.05, ZPD.offset=.02,
                curriculum.start.points=curriculum.start.points,
                curriculum.widths=curriculum.widths,
                slope2=slope2, rate=6, threshold=.00001,
                integration.points=2000)
}
}
