% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_growth_rate_function.r
\name{school.growth.rate}
\alias{school.growth.rate}
\title{Function for calculating growth resulting from school at every possible achievement level,
 given a level of achievement, ZPD parameters, and school curriculum parameters.
 This is used to populate the lookup table for school growth.
Called by build.school.lookup()}
\usage{
school.growth.rate(
  achievement,
  ZPD.offset,
  ZPD.width,
  slope1,
  slope2,
  start,
  end,
  integration.points
)
}
\arguments{
\item{achievement}{a vector of achievement levels}

\item{ZPD.offset}{offset of the student's ZPD from the current level of achievement}

\item{ZPD.width}{the radius of the student's ZPD}

\item{slope1}{the slope of the lower leg of the school curriculum}

\item{slope2}{the slope of the upper leg of the school curriculum}

\item{start}{the curriculum starting location}

\item{end}{the curriculum ending location}

\item{integration.points}{integration points for numerical integration}
}
\description{
Function for calculating growth resulting from school at every possible achievement level,
 given a level of achievement, ZPD parameters, and school curriculum parameters.
 This is used to populate the lookup table for school growth.
Called by build.school.lookup()
}
\examples{
\dontrun{
school.growth.rate(achievement=seq(0,1, .1), integration.points=200, ZPD.offset=.05,
  ZPD.width=.04, slope1=10, slope2=20, start=.1, end=.2)
  }

}
