% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_growth_function.r
\name{school.growth}
\alias{school.growth}
\title{Function for calculating the baseline growth due to school exposure during interval t
called by school.growth()}
\usage{
school.growth(
  x,
  achievement,
  ZPD.offset,
  ZPD.width,
  slope1,
  slope2,
  start,
  end
)
}
\arguments{
\item{x}{achievement value}

\item{achievement}{the student's current achievement}

\item{ZPD.offset}{offset of the student's ZPD from the current level of achievement}

\item{ZPD.width}{the radius of the student's ZPD}

\item{slope1}{the slope of the lower leg of the school curriculum}

\item{slope2}{the slope of the upper leg of the school curriculum}

\item{start}{the curriculum starting location}

\item{end}{the curriculum ending location}
}
\description{
Function for calculating the baseline growth due to school exposure during interval t
called by school.growth()
}
\examples{
\dontrun{
school.growth(x=.25, achievement=.22, ZPD.offset=0, ZPD.width=.03, slope1=30,
slope2=30, start=.2, end=.25)
}

}
