% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeContext_function.r
\name{visualizeContext}
\alias{visualizeContext}
\title{Function for visualizing the learning context, including the school and (optionally) home.}
\usage{
visualizeContext(
  home.learning.decay.rate = NULL,
  curriculum.start.points,
  curriculum.widths,
  curriculum.review.slopes,
  curriculum.advanced.slopes,
  assignment,
  points = 500,
  annotate = FALSE,
  schoolcolor = "gray25",
  homecolor = "gray25",
  linecolor = "indianred2",
  versionlabels = NULL,
  zoomschool = "truncate"
)
}
\arguments{
\item{home.learning.decay.rate}{Scalar, An optional value specifying the exponential decay parameter for the
home learning function. If provided, the background will be shaded in proportion to the intensity of
home learning. Larger values indicate more rapid drop-off. Must be greater than 1. If omitted, only the
school curricular context is shown. The \code{visualizHome()} can also be used to visualize, understand, and
select appropriate values. Defaults to NULL.}

\item{curriculum.start.points}{a matrix or list of matrices providing the start points of the school curriculum to
be offered during each grade (or other time division). The matrices should have one column and one row per grade. Each
list entry corresponds to a version of the curriculum, such as 'typical' or 'remedial.' The
\code{visualizeSchool()} function can also be used to visualize, understand, and select appropriate
values.}

\item{curriculum.widths}{a matrix or list of matrices providing the widths or spans of the school curriculum to
be offered during each grade (or other time division). The matrices should have one column and one row per grade. Each
list entry corresponds to a version of the curriculum, such as 'typical' or 'remedial.' The
\code{visualizeSchool()} function can also be used to visualize, understand, and select appropriate
values.}

\item{curriculum.review.slopes}{A matrix or list of matrices describing the steepness of the school curriculum cutoff
at the lower range. Conceptually controls the amount of review content. Small numbers indicate a shallower slope and
therefore more review content. As general guidance, values in the range of 10-20 describe heavy review, while 50-100
decribes little review, though this depends on the scale of the curriculum. The
\code{visualizeSchool()} function can also be used to visualize, understand, and select appropriate
values.}

\item{curriculum.advanced.slopes}{a matrix or list of matrices describing the steepness of the school curriculum
cutoff at the upper range. Conceptually controls the amount of advanced content. Small numbers indicate a shallower slope and
therefore more advanced content. As general guidance, values in the range of 10-20 describe heavy review, while 50-100
decribes little review, though this depends on the scale of the curriculum. The
\code{visualizeSchool()} function can also be used to visualize, understand, and select appropriate
values.}

\item{assignment}{A vector of integer values assigning school curricula to time intervals. The length is the number of
time intervals to simulate. Each entry contains a number representing which grade-level curriculum to present. Zero
denotes periods of no school instruction, such as summer breaks. The numbers correspond to the row index of the
 \code{curriculum.start.points}, \code{curriculum.widths}, \code{curriculum.review.slopes}, and
 \code{curriculum.advanced.slopes} objects.}

\item{points}{Integer value desribing the number of points within the span of the curriculum
for which the intensity is calculated. This controls the smoothness of the gradient of intensity
shading in the plot. Larger values yield better smoothness but require longer execution time.
Defaults to 500.}

\item{annotate}{logical. should the figure be annotated with the row numbers of the school
curricula? defaults to FALSE}

\item{schoolcolor}{The color used to represent the school curricular intensity. Defaults to "grey25".}

\item{homecolor}{The color used to represent the home curricular intensity. Defaults to "grey25". Has an effect
only if a non-null value is given for \code{rate}.}

\item{linecolor}{The color used for the horizontal reference lines representing where the
curriculum has full intensity. Defaults to "indianred2".}

\item{versionlabels}{A character vector providing labels for the facets representing the different
versions of the curriculum (e.g., typical, remedial).}

\item{zoomschool}{One of "none", "school", or "truncate". Sets the range of the y-axis of the plot when a non-null
value is provided for \code{rate}. "none" does not zoom. "school" shows achievement of zero through the maximum
level of the school curriculum. "truncate" shows the minimum and maximum achievement levels of the school
curriculum.}

\item{rate}{Scalar, the exponential decay parameter describing the home curriculum function. If a value is
given, the background is shaded proportional to the home learning rate. If NULL, only the school
curriculum is displayed. Defaults to NULL.}
}
\description{
Function for visualizing the learning context, including the school and (optionally) home.
}
\examples{
#' # assignment object simulating starting kindergarten on time 201
#  Kindergarten for 200 days, followed by 100 days of summer
#  then 200 days of first grade
assignment <- c(rep(0, times=200), rep(1, times=200),
                rep(0, times=100), rep(2, times=200))

# define school curriculum
curriculum.start.points <- list(
    # "typical curriculum" start points for K and first grade
  matrix(c(.2, .26), nrow=2, ncol=1),
    # "advanced curriculum" start points for K and first grade
  matrix(c(.22, .29), nrow=2, ncol=1)
)

curriculum.widths <- list(
  # "typical curriculum" widths for K and first grade
  matrix(c(.03, .03), nrow=2, ncol=1),
  # "advanced curriculum" widths for K and first grade
  matrix(c(.04, .04), nrow=2, ncol=1)
)

curriculum.review.slopes <- list(
  # "typical curriculum" review slopes for K and first grade
  matrix(c(15, 15), nrow=2, ncol=1),
  # "advanced curriculum" review slopes for K and first grade
  matrix(c(30, 30), nrow=2, ncol=1)
  )

curriculum.advanced.slopes <- list(
  # "typical curriculum" advanced slopes for K and first grade
  matrix(c(50, 50), nrow=2, ncol=1),
  # "advanced curriculum" advanced slopes for K and first grade
  matrix(c(25, 25), nrow=2, ncol=1)
  )

visualizeContext(home.learning.decay.rate=6,
                  curriculum.start.points=curriculum.start.points,
                  curriculum.widths=curriculum.widths,
                  curriculum.review.slopes=curriculum.review.slopes,
                  curriculum.advanced.slopes=curriculum.advanced.slopes,
                  assignment=assignment,
                  annotate=TRUE,
                  versionlabels = c("Typical", "Advanced"),
                  linecolor="blue",
                  zoomschool="truncate")

}
\seealso{
[visualizeZPD()], [visualizeSchool()], [visualizeHome()], and [visualizeTrajectories()]

Other visualizations: 
\code{\link{visualizeHome}()},
\code{\link{visualizeSchool}()},
\code{\link{visualizeTrajectories}()}
}
\concept{visualizations}
