% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeHome_function.r
\name{visualizeHome}
\alias{visualizeHome}
\title{Function for visualizing the home curriculum.}
\usage{
visualizeHome(rate, max = 1)
}
\arguments{
\item{rate}{Scalar decay rate of the exponential function describing the home curriulum}

\item{max}{The maximum value to plot on the x-axis; defaults to 1}
}
\description{
\code{visualizeHome} is a function for visualizing the home curriculum function.
}
\details{
\code{visualizeHome} plots the home curriculum. It is useful for exploring
   the consequences of the \code{rate} argument to \code{ZPDGrowthTrajectories()}. The function returns a
   \code{ggplot} object that can be modified with typical \code{ggplot2} arguments.
}
\examples{
visualizeHome(rate=5)
}
