% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeHomeCurriculum_function.R
\name{visualizeHomeCurriculum}
\alias{visualizeHomeCurriculum}
\title{Function for visualizing the home curriculum.}
\usage{
visualizeHomeCurriculum(shape1 = 1, shape2)
}
\arguments{
\item{shape1}{The \code{shape1} argument to the \code{dbeta} function for describing the shape of the
home curriculum. Defaults to 1.0.}

\item{shape2}{The \code{shape2} argument to the \code{dbeta} function for describing the shape of the
home curriculum.}
}
\description{
\code{visualizeHomeCurriculum} is a function for visualizing the home curriculum function.
}
\details{
\code{visualizeHomeCurriculum} plots the home curriculum. It is useful for exploring
   the consequences of the \code{home.curriculum.shape1} and \code{home.curriculum.shape2}
   arguments to \code{ZPDGrowthTrajectories()} and related functions. The function returns a 
   \code{ggplot} object that can be modified with typical \code{ggplot2} arguments.
}
