% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeLearning_function.r
\name{visualizeLearning}
\alias{visualizeLearning}
\title{Function for visualizing the learning context, including the school and (optionally) home.}
\usage{
visualizeLearning(
  curriculum.start.points,
  curriculum.widths,
  assignment,
  slope1,
  slope2,
  points = 500,
  annotate = FALSE,
  schoolcolor = "gray25",
  homecolor = "gray25",
  linecolor = "indianred2",
  versionlabels = NULL,
  rate = NULL,
  zoomschool = "truncate"
)
}
\arguments{
\item{curriculum.start.points}{a matrix or list of matrices providing the start points of each
grade level (rows) and version (columns) of the school curriculum.}

\item{curriculum.widths}{a matrix or list of matrices providing the spans of each grade level
(rows) and version (columns) of the school curriculum.}

\item{assignment}{a vector. The length is the number of time intervals to simulate. Each entry contains a number representing which
grade-level curriculum to present. Zero denotes summers. The numbers correspond to the row index of the
 \code{curriculum.start.points} and \code{curriculum.widths} objects.}

\item{slope1}{The steepness of the school curriculum cutoff at the lower range. Conceptually controls the amount of review content.}

\item{slope2}{The slope of the school curriculum at the upper range. Conceptually controls the amount of advanced content.}

\item{points}{Integer value desribing the number of points within the span of the curriculum
for which the intensity is calculated. This controls the smoothness of the gradient of intensity
shading in the plot. Larger values yield better smoothness but require longer execution time.
Defaults to 500.}

\item{annotate}{logical. should the figure be annotated with the row numbers of the school
curricula? defaults to FALSE}

\item{schoolcolor}{The color used to represent the school curricular intensity. Defaults to "grey25".}

\item{homecolor}{The color used to represent the home curricular intensity. Defaults to "grey25". Has an effect
only if a non-null value is given for \code{rate}.}

\item{linecolor}{The color used for the horizontal reference lines representing where the
curriculum has full intensity. Defaults to "indianred2".}

\item{versionlabels}{A character vector providing labels for the facets representing the different
versions of the curriculum (e.g., typical, remedial).}

\item{rate}{Scalar, the exponential decay parameter describing the home curriculum function. If a value is
given, the background is shaded proportional to the home learning rate. If NULL, only the school
curriculum is displayed. Defaults to NULL.}

\item{zoomschool}{One of "none", "school", or "truncate". Sets the range of the y-axis of the plot when a non-null
value is provided for \code{rate}. "none" does not zoom. "school" shows achievement of zero through the maximum
level of the school curriculum. "truncate" shows the minimum and maximum achievement levels of the school
curriculum.}
}
\description{
Function for visualizing the learning context, including the school and (optionally) home.
}
\examples{
#' # assignment object simulating starting kindergarten on time 201
#  Kindergarten for 200 days, followed by 100 days of summer
#  then 200 days of first grade
assignment <- c(rep(0, times=200), rep(1, times=200),
                rep(0, times=100), rep(2, times=200))

# define school curriculum
curriculum.start.points <- list(
    # "typical curriculum" start points for K and first grade
  matrix(c(.2, .26), nrow=2, ncol=1),
    # "advanced curriculum" start points for K and first grade
  matrix(c(.22, .29), nrow=2, ncol=1)
)

curriculum.widths <- list(
  # "typical curriculum" widths for K and first grade
  matrix(c(.03, .03), nrow=2, ncol=1),
  # "advanced curriculum" widths for K and first grade
  matrix(c(.04, .04), nrow=2, ncol=1)
)


}
