% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeSchool_function.r
\name{visualizeSchool}
\alias{visualizeSchool}
\title{Function for visualizing the school curriculum.}
\usage{
visualizeSchool(start.point, width, slope1, slope2)
}
\arguments{
\item{start.point}{The point at which the curriculum reaches full intensity. This value would become one of
the elements of the curriculum.start.points argument to \code{ZPDGrowthTrajectories}.}

\item{width}{The highest point at which the curriculum is at full intensity. This value would become one of
the elements of the curriculum.widths argument to \code{ZPDGrowthTrajectories}.}

\item{slope1}{The steepness of the school curriculum cutoff at the lower range.
Conceptually controls the amount of review content.}

\item{slope2}{The slope of the school curriculum at the upper range.
Conceptually controls the amount of advanced content.}
}
\description{
\code{visualizeSchool} plots school curriculum. It is useful for exploring
   the consequences of the \code{curriculum.start.points}, \code{curriculum.widths},
   \code{curriculum.lower.slope}, \code{curriculum.upper.slope}, and \code{alpha} arguments to
   \code{ZPDGrowthTrajectories()}. The function returns a
   \code{ggplot} object that can be modified with typical \code{ggplot2} arguments.
}
\examples{
visualizeSchool(start.point=.2, width=.1, slope1=10, slope2=15)

}
