% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeSchoolCurriculum_function.r
\name{visualizeSchoolCurriculum}
\alias{visualizeSchoolCurriculum}
\title{Function for visualizing the school curriculum.}
\usage{
visualizeSchoolCurriculum(start.point, width, lower.slope = 8,
  upper.slope = 300, alpha = 1)
}
\arguments{
\item{start.point}{a scalar value between 0 and 1 providing the lower bound value of achievement
at which the school curriculum has maximum intensity.}

\item{width}{a scalar value between 0 and 1 providing the upper bound value of achievement
at which the school curriculum has maximum intensity.}

\item{lower.slope}{scalar argument to the \code{mode1} argument of 
\code{trapedoid::dtrapezoid}; controls steepness of the school curriculum cutoff at the 
lower range. Conceptually controls the amount of review content. Defaults to 8.}

\item{upper.slope}{scalar argument to the \code{mode2} argument of 
\code{trapedoid::dtrapezoid}; controls steepness of the school curriculum cutoff at the 
upper range. Conceptually controls the amount of advanced content. Defaults to 300.}

\item{alpha}{scalar argument to the \code{mode2} argument of \code{trapedoid::dtrapezoid}. 
Controls the shape of the school curriculum at maximum intensity. Defaults to 1.}
}
\description{
\code{visualizeSchoolCurriculum} is a function for visualizing the school curriculum.
}
\details{
\code{visualizeSchoolCurriculum} plots school curriculum. It is useful for exploring
   the consequences of the \code{curriculum.start.points}, \code{curriculum.widths}, 
   \code{curriculum.lower.slope}, \code{curriculum.upper.slope}, and \code{alpha} arguments to 
   \code{ZPDGrowthTrajectories()}. The function returns a 
   \code{ggplot} object that can be modified with typical \code{ggplot2} arguments.
}
