% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeZPD_function.R
\name{visualizeZPD}
\alias{visualizeZPD}
\title{Visualize the zone of proximal development}
\usage{
visualizeZPD(achievement = 0.3, zpd.sd, zpd.offset)
}
\description{
\code{visualizeZPD} is a function for visualizing the zone of proximal development (ZPD).
}
\details{
\code{visualizeZPD} plots the ZPD against the achievement. It is useful for exploring
   the consequences of the \code{zpd.sd} and \code{zpd.offset} arguments to 
   \code{ZPDGrowthTrajectories()} and related functions. The function returns a 
   \code{ggplot} object that can be modified with typical \code{ggplot2} arguments.

@param achievement A scalar value between 0 and 1 indicating the current achievement level.
  Defaults to 0.3. 
 @param zpd.sd A positive scalar describing the standard deviation of the normal curve describing
   the zone of proximal development. Controls the 'width' of the ZPD.
 @param zpd.offset A scalar (typically positive) describing the distance between achievement and 
   the peak intensity of the ZPD.
}
