% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_function.R
\name{ZPDGrowthTrajectories}
\alias{ZPDGrowthTrajectories}
\title{Function for creating simulated growth trajectories from the theoretical model.}
\usage{
ZPDGrowthTrajectories(output.format = "wide", days, assignment,
  curriculum.start.points, curriculum.widths, dosage, learning.rates,
  decay.rates, initial.achievements, home.environments,
  integration.points = 200, curriculum.lower.slope = 8,
  curriculum.upper.slope = 300, alpha = 1, home.curriculum.shape1 = 1,
  home.curriculum.shape2 = 4.6, zpd.offset = 0.022, zpd.sd = 0.042,
  zpd.scale = 0.055, decay.weight = 0.005, useGPU = FALSE)
}
\arguments{
\item{output.format}{Format of the results, "long" for long format, "wide" for wide. Defaults to "wide".}

\item{days}{the number of days to simulate.}

\item{assignment}{a vector. The length is the number of days to simulate. Each entry contains a number representing which
grade-level curriculum to present. Zero denotes summers. The numbers correspond to the row index of the
 \code{curriculum.start.points} and \code{width} objects.}

\item{curriculum.start.points}{a matrix providing the start points of each grade level (rows) and version (columns) of the
school curriculum.}

\item{curriculum.widths}{a matrix providing the spans of each grade level (rows) and version (columns) of the
school curriculum.}

\item{dosage}{scalar dose parameter, controls mixing of school curriculum and home curriculum during school years.
Must be [0,1].}

\item{learning.rates}{vector of learning rates, one per child.}

\item{decay.rates}{vector of decay rates, one per child. Controls forgetting.}

\item{initial.achievements}{vector of initial achievements, one per child.}

\item{home.environments}{vector of home environments, one per child.}

\item{integration.points}{number of integration points. Controls tradeoff between accuracy and execution speed.
Defaults to 200.}

\item{curriculum.lower.slope}{scalar argument to the \code{mode1} argument of \code{trapedoid::dtrapezoid}; controls
steepness of the school curriculum cutoff at the lower range. Conceptually controls the amount of review content.}

\item{curriculum.upper.slope}{scalar argument to the \code{mode2} argument of \code{trapedoid::dtrapezoid}; controls
steepness of the school curriculum cutoff at the upper range. Conceptually controls the amount of advanced content.}

\item{alpha}{scalar argument to the \code{mode2} argument of \code{trapedoid::dtrapezoid}. Defaults to 1.}

\item{home.curriculum.shape1}{The \code{shape1} argument to the \code{dbeta} function for describing the shape of the
home curriculum. Defaults to 1.}

\item{home.curriculum.shape2}{The \code{shape2} argument to the \code{dbeta} function for describing the shape of the
home curriculum. Defaults to 4.6.}

\item{zpd.offset}{scalar value, measured on the same scale as achievement, describing where the ZPD peaks
relative to the current achievement. Defaults to 0.022.}

\item{zpd.sd}{the standard deviation of the ZPD function, controls its width. Defaults to 0.042.}

\item{zpd.scale}{scaling factor for the ZPD function, controls overall learning rate. Default is 0.055.}

\item{decay.weight}{scalar parameter, global control of decay rates.}

\item{useGPU}{should the code be executed on the GPU? May offer performance boost on large simulations. The \code{gpuR}
package must be installed. Defaults to FALSE.}
}
\description{
description.
}
