% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeTrajectories_function.r
\name{describeTrajectories}
\alias{describeTrajectories}
\title{Create summary statistics for ZPDGrowthTrajectories}
\usage{
describeTrajectories(
  trajectories,
  byTransition = TRUE,
  byVersion = TRUE,
  times = NULL
)
}
\arguments{
\item{trajectories}{An object of class \code{ZPD} produced by the \code{ZPDGrowthTrajectories()}
function. If needed, this object will be converted internally to "long" format.}

\item{byTransition}{Logical. Should descriptives be provided only at curricular transitions (as
well as the first and last time point)? Defaults to TRUE.}

\item{byVersion}{Logical. Should descriptives be broken down by version of the curriculum (
e.g., typical, remedial)? Defaults to TRUE.}

\item{times}{Optional. A vector of specific times at which descriptives should be computed.
Defaults to NULL.}
}
\value{
An object of class \code{tibble}
}
\description{
\code{describeTrajectories} calculates summary statistics by time interval and version of the
  school curriculum for for \code{ZPDGrowthTrajectories()} output.
}
\seealso{
\code{\link{visualizeTrajectories}} for plotting the trajectories.

\code{\link{ZPDGrowthTrajectories}} for simulating growth trajectories.
}
