% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeTrajectories_function.r
\name{describeTrajectories}
\alias{describeTrajectories}
\title{Create summary statistics for ZPDGrowthTrajectories}
\usage{
describeTrajectories(
  trajectories,
  byTransition = TRUE,
  byVersion = TRUE,
  times = NULL
)
}
\arguments{
\item{trajectories}{An object of class \code{ZPD} produced by the \code{ZPDGrowthTrajectories()}
function. If needed, this object will be converted internally to "long" format.}

\item{byTransition}{Logical. Should descriptives be provided only at curricular transitions (as
well as the first and last time point)? Defaults to TRUE.}

\item{byVersion}{Logical. Should descriptives be broken down by version of the curriculum (
e.g., typical, remedial)? Defaults to TRUE.}

\item{times}{Optional. A vector of specific times at which descriptives should be computed.
Defaults to NULL.}
}
\value{
An object of class \code{tibble}
}
\description{
\code{describeTrajectories} calculates summary statistics by time interval and version of the
  school curriculum for for \code{ZPDGrowthTrajectories()} output.
}
\examples{

# learning rate
learn.rate <- c(.08, .10, .12, .18)

# decay rate
decay.rate <- c(.04, .03, .02, .01)

# initial achievement
initial.ach <- rep(0, times=4)

# quality of home environment
home.env <- c(.06, .12, .15, .20)

# assignment object simulating starting kindergarten on time 801
#  Kindergarten for 200 days, followed by 100 days of summer
#  then 200 days of first grade
assignment <- c(rep(0, times=800), rep(1, times=200),
                rep(0, times=100), rep(2, times=200))

# define school curriculum
curriculum.start.points <- list(
    # "typical curriculum" start points for K and first grade
  matrix(c(.2, .26), nrow=2, ncol=1),
    # "advanced curriculum" start points for K and first grade
  matrix(c(.22, .29), nrow=2, ncol=1)
)

curriculum.widths <- list(
  # "typical curriculum" widths for K and first grade
  matrix(c(.04, .04), nrow=2, ncol=1),
  # "advanced curriculum" widths for K and first grade
  matrix(c(.05, .05), nrow=2, ncol=1)
)

curriculum.review.slopes <- list(
  # "typical curriculum" review slopes for K and first grade
  matrix(c(30, 30), nrow=2, ncol=1),
  # "advanced curriculum" review slopes for K and first grade
  matrix(c(60, 60), nrow=2, ncol=1)
)

curriculum.advanced.slopes <- list(
  # "typical curriculum" advanced slopes for K and first grade
  matrix(c(50, 50), nrow=2, ncol=1),
  # "advanced curriculum" advanced slopes for K and first grade
  matrix(c(25, 25), nrow=2, ncol=1)
)

# students 1 and 2 get typical curriculum, 3 and 4 get advanced
which.curriculum <- c(1,1,2,2)

 y <- ZPDGrowthTrajectories(learn.rate=learn.rate, home.env=home.env,
                        decay.rate=decay.rate, initial.ach=initial.ach,
                        ZPD.width=.05, ZPD.offset=.02,
                        home.learning.decay.rate=6,
                        curriculum.start.points=curriculum.start.points,
                        curriculum.widths=curriculum.widths,
                        curriculum.review.slopes=curriculum.review.slopes,
                        curriculum.advanced.slopes=curriculum.advanced.slopes,
                        assignment=assignment, dosage=.8,
                        adaptive.curriculum=FALSE,
                        which.curriculum=which.curriculum,
                        school.weight=.5, home.weight=1, decay.weight=.05,
                        verbose=TRUE)
describeTrajectories(y, byVersion=FALSE)

describeTrajectories(y, byTransition=FALSE, times=c(100, 300, 650))

}
\seealso{
\code{\link{visualizeTrajectories}} for plotting the trajectories.

\code{\link{ZPDGrowthTrajectories}} for simulating growth trajectories.
}
