% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeHome_function.r
\name{visualizeHome}
\alias{visualizeHome}
\title{Function for visualizing the home curriculum.}
\usage{
visualizeHome(home.learning.decay.rate, max = 1)
}
\arguments{
\item{home.learning.decay.rate}{Scalar, the exponential decay parameter for the home learning function.
Larger values indicate a more rapid drop-off. Must be greater than 1. The functions \code{visualizHome()}
and \code{visualizeContext()} can be used to visualize, understand, and select appropriate values.}

\item{max}{The maximum value to plot on the x-axis; defaults to 1}
}
\value{
An object of class \code{ggplot2}
}
\description{
\code{visualizeHome} is a function for visualizing the home curriculum function.
}
\details{
\code{visualizeHome} plots the home curriculum. It is useful for exploring
   the consequences of the \code{rate} argument to \code{ZPDGrowthTrajectories()}. The function returns a
   \code{ggplot} object that can be modified with typical \code{ggplot2} arguments.
}
\examples{
visualizeHome(home.learning.decay.rate=5)
}
\seealso{
Other visualizations: 
\code{\link{visualizeContext}()},
\code{\link{visualizeSchool}()},
\code{\link{visualizeTrajectories}()},
\code{\link{visualizeZPD}()}
}
\concept{visualizations}
