% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeSchool_function.r
\name{visualizeSchool}
\alias{visualizeSchool}
\title{Function for visualizing the school curriculum.}
\usage{
visualizeSchool(start.point, width, review.slope, advanced.slope)
}
\arguments{
\item{start.point}{The point at which the curriculum reaches full intensity. This value would become one of
the elements of the \code{curriculum.start.points} argument to \code{ZPDGrowthTrajectories()}.}

\item{width}{The highest point at which the curriculum is at full intensity. This value would become one of
the elements of the \code{curriculum.widths} argument to \code{ZPDGrowthTrajectories()}.}

\item{review.slope}{The steepness of the school curriculum cutoff at the lower range.
 Conceptually controls the amount of review content. This value would become one of
the elements of the \code{curriculum.review.slopes} argument to \code{ZPDGrowthTrajectories()}.}

\item{advanced.slope}{The slope of the school curriculum at the upper range.
 Conceptually controls the amount of advanced content. This value would become one of
the elements of the \code{curriculum.advanced.slopes} argument to \code{ZPDGrowthTrajectories()}.}
}
\value{
An object of class \code{ggplot2}
}
\description{
\code{visualizeSchool()} is a function for visualizing the school curriculum function.
}
\details{
\code{visualizeSchool()} plots the school curriculum. It is useful for exploring
   the consequences of the \code{curriculum.start.points}, \code{curriculum.widths},
   \code{curriculum.lower.slopes}, \code{curriculum.upper.slopes}, and \code{alpha} arguments to
   \code{ZPDGrowthTrajectories()}. The function returns a
   \code{ggplot} object that can be modified with typical \code{ggplot2} arguments.
}
\examples{
visualizeSchool(start.point=.2, width=.1, review.slope=10, advanced.slope=15)

}
\seealso{
Other visualizations: 
\code{\link{visualizeContext}()},
\code{\link{visualizeHome}()},
\code{\link{visualizeTrajectories}()},
\code{\link{visualizeZPD}()}
}
\concept{visualizations}
