% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeTrajectories_function.r
\name{visualizeTrajectories}
\alias{visualizeTrajectories}
\title{Function for visualizing trajectories generated by ZPDGrowthTrajectories}
\usage{
visualizeTrajectories(trajectories, showTransitions = FALSE)
}
\arguments{
\item{trajectories}{An object of class \code{ZPD} produced by the \code{ZPDGrowthTrajectories()}
function. If needed, this object be converted internally to "long" format suitable for
\code{ggplot}.The function returns a \code{ggplot} object that can be modified with typical
\code{ggplot2} arguments.}

\item{showTransitions}{Logical. Should vertical lines be drawn when the school curriculum changes?
Defaults to FALSE}
}
\description{
\code{visualizeTrajectories} plots synthetic growth trajectories using \code{ggplot2}.
}
\details{
\code{visualizeTrajectories} plots the trajectories generated by the \code{ZPDGrowthTrajectories()}
 function. The figure plots achievement versus time and represents each student's trajectory using a
 different colored line.
}
\seealso{
\code{\link{describeTrajectories}} for calculating summary statistics of the generated
 trajectories.

\code{\link{ZPDGrowthTrajectories}} for simulating growth trajectories.

Other visualizations: 
\code{\link{visualizeContext}()},
\code{\link{visualizeHome}()},
\code{\link{visualizeSchool}()},
\code{\link{visualizeZPD}()}
}
\concept{visualizations}
