% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeTrajectories_function.r
\name{visualizeTrajectories}
\alias{visualizeTrajectories}
\title{Function for visualizing trajectories generated by ZPDGrowthTrajectories}
\usage{
visualizeTrajectories(
  trajectories,
  showTransitions = FALSE,
  timerange = NULL,
  version = NULL
)
}
\arguments{
\item{trajectories}{An object of class \code{ZPD} produced by the \code{ZPDGrowthTrajectories()}
function. If needed, this object be converted internally to "long" format suitable for
\code{ggplot}.The function returns a \code{ggplot} object that can be modified with typical
\code{ggplot2} arguments.}

\item{showTransitions}{Logical. Should vertical lines be drawn when the school curriculum changes?
Defaults to FALSE}

\item{timerange}{Optional numeric vector providing limits for the x axis. Defaults to NULL.}

\item{version}{Optional numeric vector for plotting trajectories only for specific version(s) of
the curriculum. For example, \code{c(1,2)} indicates that only trajectories for versions 1 and 2
should be plotted}
}
\description{
\code{visualizeTrajectories} plots synthetic growth trajectories using \code{ggplot2}.
}
\details{
\code{visualizeTrajectories} plots the trajectories generated by the \code{ZPDGrowthTrajectories()}
 function. The figure plots achievement versus time and represents each student's trajectory using a
 different colored line.
}
\examples{
# learning rate
learn.rate <- c(.08, .10, .12, .18)

# decay rate
decay.rate <- c(.04, .03, .02, .01)

# initial achievement
initial.ach <- rep(0, times=4)

# quality of home environment
home.env <- c(.06, .12, .15, .20)

# assignment object simulating starting kindergarten on time 201
#  Kindergarten for 200 days, followed by 100 days of summer
#  then 200 days of first grade
assignment <- c(rep(0, times=200), rep(1, times=200),
                rep(0, times=100), rep(2, times=200))

# define school curriculum
curriculum.start.points <- list(
    # "typical curriculum" start points for K and first grade
  matrix(c(.2, .26), nrow=2, ncol=1),
    # "advanced curriculum" start points for K and first grade
  matrix(c(.22, .29), nrow=2, ncol=1)
)

curriculum.widths <- list(
  # "typical curriculum" widths for K and first grade
  matrix(c(.03, .03), nrow=2, ncol=1),
  # "advanced curriculum" widths for K and first grade
  matrix(c(.04, .04), nrow=2, ncol=1)
)

curriculum.review.slopes <- list(
  # "typical curriculum" review slopes for K and first grade
  matrix(c(15, 15), nrow=2, ncol=1),
  # "advanced curriculum" review slopes for K and first grade
  matrix(c(30, 30), nrow=2, ncol=1)
  )

curriculum.advanced.slopes <- list(
  # "typical curriculum" advanced slopes for K and first grade
  matrix(c(50, 50), nrow=2, ncol=1),
  # "advanced curriculum" advanced slopes for K and first grade
  matrix(c(25, 25), nrow=2, ncol=1)
  )

 # students 1 and 2 get typical curriculum, 3 and 4 get advanced
 which.curriculum <- c(1,1,2,2)

 y <- ZPDGrowthTrajectories(learn.rate=learn.rate, home.env=home.env,
                        decay.rate=decay.rate, initial.ach=initial.ach,
                        ZPD.width=.05, ZPD.offset=.02,
                        home.learning.decay.rate=6,
                        curriculum.start.points=curriculum.start.points,
                        curriculum.widths=curriculum.widths,
                        curriculum.review.slopes=curriculum.review.slopes,
                        curriculum.advanced.slopes=curriculum.advanced.slopes,
                        assignment=assignment, dosage=.8,
                        adaptive.curriculum=FALSE,
                        which.curriculum=which.curriculum,
                        school.weight=.5, home.weight=1, decay.weight=.05,
                        verbose=TRUE)
visualizeTrajectories(y)

visualizeTrajectories(y, timerange=c(500, 700))


}
\seealso{
\code{\link{describeTrajectories}} for calculating summary statistics of the generated
 trajectories.

\code{\link{ZPDGrowthTrajectories}} for simulating growth trajectories.

Other visualizations: 
\code{\link{visualizeContext}()},
\code{\link{visualizeHome}()},
\code{\link{visualizeSchool}()},
\code{\link{visualizeZPD}()}
}
\concept{visualizations}
