% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeTrajectories_function.R
\name{visualizeTrajectories}
\alias{visualizeTrajectories}
\title{Visualize growth trajectories}
\usage{
visualizeTrajectories(trajectories)
}
\arguments{
\item{trajectories}{A data frame produced by the \code{ZPDGrowthTrajectories()}, 
\code{ZPDGrowthTrajectories_studentmatrix()}, or \code{ZPDGrowthTrajectories_multicore()} functions.
Will be converted internally to "long" format suitable for \code{ggplot}. The function returns a 
\code{ggplot} object that can be modified with typical \code{ggplot2} arguments.}
}
\description{
Function for visualizing synthethic growth trajectories.
#' 
\code{visualizeTrajectories} plots synthetic growth trajectories using \code{ggplot2}.
}
