% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeZPD_function.r
\name{visualizeZPD}
\alias{visualizeZPD}
\title{Visualize the zone of proximal development}
\usage{
visualizeZPD(achievement, zpd.offset, zpd.width)
}
\arguments{
\item{achievement}{the current level of achievement, defaults to 0.3}

\item{zpd.offset}{scalar value, measured on the same scale as achievement, describing where the ZPD peaks
relative to the current achievement.}

\item{zpd.width}{the radius of the ZPD.}
}
\value{
An object of class \code{ggplot2}
}
\description{
\code{visualizeZPD} is a function for visualizing the zone of proximal development (ZPD).
}
\details{
\code{visualizeZPD} plots the ZPD against the achievement. It is useful for exploring
   the consequences of the \code{zpd.width} and \code{zpd.offset} arguments to
   \code{ZPDGrowthTrajectories()}. The function returns a \code{ggplot} object that
   can be modified with typical \code{ggplot2} arguments.
}
\examples{
visualizeZPD(achievement=.3, zpd.offset=.02, zpd.width=.06)
}
\seealso{
Other visualizations: 
\code{\link{visualizeContext}()},
\code{\link{visualizeHome}()},
\code{\link{visualizeSchool}()},
\code{\link{visualizeTrajectories}()}
}
\concept{visualizations}
