% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZPDGrowthTrajectories_function.r
\name{ZPDGrowthTrajectories}
\alias{ZPDGrowthTrajectories}
\title{Function for creating simulated growth trajectories from the theoretical model.}
\usage{
ZPDGrowthTrajectories(
  learn.rate,
  home.env,
  decay.rate,
  initial.ach,
  school.weight,
  home.weight,
  decay.weight,
  dosage,
  assignment,
  adaptive.curriculum = FALSE,
  which.curriculum = NULL,
  ZPD.width,
  ZPD.offset,
  curriculum.start.points,
  curriculum.widths,
  slope1,
  slope2,
  rate = 6,
  integration.points = 250,
  threshold = 1e-05,
  verbose = TRUE,
  output.format = "long"
)
}
\arguments{
\item{learn.rate}{vector of learning rates, one per child.}

\item{home.env}{vector of home environments, one per child.}

\item{decay.rate}{vector of decay rates, one per child. Controls forgetting.}

\item{initial.ach}{vector of initial achievements, one per child.}

\item{school.weight}{a global scalar for adjusting the growth from school exposure; prevents having to adjust
all learn.rate values}

\item{home.weight}{a global scalar for adjusting the growth from home; prevents having to adjust all home.env values}

\item{decay.weight}{a global scalar for adjusting the rate of forgetting; prevents having to adjust all decay.rate values}

\item{dosage}{scalar dose parameter, controls mixing of school curriculum and home curriculum during school years.
Must be [0,1].}

\item{assignment}{a vector. The length is the number of days to simulate. Each entry contains a number representing which
grade-level curriculum to present. Zero denotes summers. The numbers correspond to the row index of the
 \code{curriculum.start.points} and \code{curriculum.widths} objects.}

\item{which.curriculum}{Defaults to NULL}

\item{curriculum.start.points}{a matrix or list of matrices providing the start points of each grade level (rows) and version (columns) of the
school curriculum.}

\item{curriculum.widths}{a matrix or list of matrices providing the spans of each grade level (rows) and version (columns) of the
school curriculum.}

\item{slope1}{The steepness of the school curriculum cutoff at the lower range. Conceptually controls the amount of review content.}

\item{slope2}{The slope of the school curriculum at the upper range. Conceptually controls the amount of advanced content.}

\item{rate}{Scalar, the exponential decay parameter describing the home curriculum function.}

\item{integration.points}{Integer number of integration points. Controls tradeoff between accuracy and execution speed.
Defaults to 200.}

\item{verbose}{logical, should status updates be printed to the console? Defaults to TRUE.}

\item{output.format}{Format of the results, "long" for long format, "wide" for wide. Defaults to "long".}

\item{days}{the number of days to simulate.}

\item{adaptive.curriculum.}{logical; if there are multiple versions of the school curriculum for each time period,
should they be assigned adaptively? TRUE means that the version that would produce the optimal achievement in
each time period is presented. FALSE means that the curriculum assignment is based on which.curriculum.
Defaults to FALSE}

\item{zpd.offset}{scalar value, measured on the same scale as achievement, describing where the ZPD peaks
relative to the current achievement.}

\item{zpd.width}{the radius of the ZPD.}
}
\description{
\code{ZPDGrowthTrajectories} creates synthetic academic achievement growth trajectories.
}
\details{
This function creates synthetic achievement growth trajectories from a quantitative intepretation
of Vygotsky theory based on an upcoming publication. The user describes
the student characteristics (learning rate, decay rate, initial achievement, and home environment) as
well as the home 'curriculum' and the school curriculum. The function will then create a simulated
growth trajectory for each student.
}
\examples{
# learning rate
learning.rates <- c(.08, .10, .12, .18)

# decay rate
decay.rates <- c(.04, .03, .02, .01)

# initial achievement
initial.achievements <- rep(0, times=4)

# quality of home environment
home.environments <- c(.06, .12, .15, .20)

# assignment object simulating starting kindergarten on day 801
#  Kindergarten for 200 days, followed by 100 days of summer
#  then 200 days of first grade
assignment <- c(rep(0, times=800), rep(1, times=200),
                rep(0, times=100), rep(2, times=200))

# define school curriculum
curriculum.start.points <- list(
    # "typical curriculum" start points for K and first grade
  matrix(c(.2, .26), nrow=2, ncol=1),
    # "advanced curriculum" start points for K and first grade
  matrix(c(.22, .29), nrow=2, ncol=1)
)

curriculum.widths <- list(
  # "typical curriculum" widths for K and first grade
  matrix(c(.03, .03), nrow=2, ncol=1),
  # "advanced curriculum" widths for K and first grade
  matrix(c(.04, .04), nrow=2, ncol=1)
)

# students 1 and 2 get typical curriculum, 3 and 4 get advanced
which.curriculum <- c(1,1,2,2)

y1 <- ZPDGrowthTrajectories(learn.rate=learn.rate, home.env=home.env,
                           decay.rate=decay.rate, initial.ach=initial.ach,
                           school.weight=1, home.weight=1, decay.weight=.25,
                           dosage=.8, assignment=assignment,
                           which.curriculum=which.curriculum,
                           adaptive.curriculum=F,
                           slope1=8, slope2=15, rate=6,
                           ZPD.width=.05, ZPD.offset=.02,
                           curriculum.start.points=curriculum.start.points,
                           curriculum.widths=curriculum.widths,
                           verbose=T, output.format="long")
}
